Extract engine watermark operation
watermark is a focused overlay operation with position resolution, opacity formatting, and output probing. This moves it into engine_watermark.py while re-exporting watermark from mcp_video.engine for compatibility.

Constraint: Preserve overlay positioning, quality args, audio copy behavior, and EditResult output.

Rejected: Move overlay_video with watermark | overlay_video has separate video/video composition behavior and should be reviewed independently.

Confidence: high

Scope-risk: narrow

Directive: Keep FFmpeg overlay operations small and independently verified.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_watermark.py

Tested: /opt/homebrew/bin/python3 watermark re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine_advanced.py::TestWatermark tests/test_server.py -k 'watermark' tests/test_cli.py -k 'watermark' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
